+++ src/main/kotlin/com/cmgapps/license/reporter/HtmlReport.kt
@@ -67,7 +67,7 @@ internal class HtmlReport(
                                 }
                             }
                         }
-                        else -> pre { +(license.id.text ?: "") }
+                        else -> pre { +(license.id.text) }
                     }
                 }
             }