+++ src/test/kotlin/com/cmgapps/license/reporter/CsvReportShould.kt
@@ -24,10 +24,10 @@ class CsvReportShould {
         assertThat(
             result,
             `is`(
-                "Name,Version,MavenCoordinates,Description,License Name,License Url\r\n" +
-                    "Test lib 1,1.0,test.group:test.artifact:1.0,proper description,Apache 2.0,https://www.apache.org/licenses/LICENSE-2.0.txt\r\n" +
-                    "Test lib 1,1.0,test.group:test.artifact:1.0,proper description,MIT License,https://opensource.org/licenses/MIT\r\n" +
-                    "Test lib 2,2.3.4,group.test2:artifact:2.3.4,descriptions of lib 2,\"The Apache Software License, Version 2.0\",https://www.apache.org/licenses/LICENSE-2.0.txt\r\n"
+                "Name,Version,MavenCoordinates,Description,SPDX-License-Identifier,License Name,License Url\r\n" +
+                    "Test lib 1,1.0,test.group:test.artifact:1.0,proper description,Apache-2.0,Apache 2.0,https://www.apache.org/licenses/LICENSE-2.0.txt\r\n" +
+                    "Test lib 1,1.0,test.group:test.artifact:1.0,proper description,MIT,MIT License,https://opensource.org/licenses/MIT\r\n" +
+                    "Test lib 2,2.3.4,group.test2:artifact:2.3.4,descriptions of lib 2,Apache-2.0,\"The Apache Software License, Version 2.0\",https://www.apache.org/licenses/LICENSE-2.0.txt\r\n"
 
             )
         )
@@ -47,8 +47,8 @@ class CsvReportShould {
         assertThat(
             result,
             `is`(
-                "Name,Version,MavenCoordinates,Description,License Name,License Url\r\n" +
-                    "\"Name with a , in it\",\"version with a \n in it\",\"groupC:articfactA:version with a \n in it\",\"description with \r in it\",\"License name with a \"\" in it\",just a plain url\r\n"
+                "Name,Version,MavenCoordinates,Description,SPDX-License-Identifier,License Name,License Url\r\n" +
+                    "\"Name with a , in it\",\"version with a \n in it\",\"groupC:articfactA:version with a \n in it\",\"description with \r in it\",,\"License name with a \"\" in it\",just a plain url\r\n"
             )
         )
     }