+++ src/test/kotlin/com/cmgapps/license/reporter/XmlReportShould.kt
@@ -30,12 +30,12 @@ class XmlReportShould {
                           proper description
                         </description>
                         <licenses>
-                          <license url="https://www.apache.org/licenses/LICENSE-2.0.txt">
+                          <license spdx-license-identifier="Apache-2.0" url="https://www.apache.org/licenses/LICENSE-2.0.txt">
                             <name>
                               Apache 2.0
                             </name>
                           </license>
-                          <license url="https://opensource.org/licenses/MIT">
+                          <license spdx-license-identifier="MIT" url="https://opensource.org/licenses/MIT">
                             <name>
                               MIT License
                             </name>
@@ -50,7 +50,7 @@ class XmlReportShould {
                           descriptions of lib 2
                         </description>
                         <licenses>
-                          <license url="https://www.apache.org/licenses/LICENSE-2.0.txt">
+                          <license spdx-license-identifier="Apache-2.0" url="https://www.apache.org/licenses/LICENSE-2.0.txt">
                             <name>
                               The Apache Software License, Version 2.0
                             </name>